+++ packages/react/src/experimental/AiChat/components/AssistantMessage.tsx
@@ -9,42 +9,38 @@ import {
 import { useI18n } from "@/lib/providers/i18n"
 import { cn } from "@/lib/utils"
 import { Markdown, type AssistantMessageProps } from "@copilotkit/react-ui"
-import { type TextMessage } from "@copilotkit/runtime-client-gql"
 import { useCallback, useRef, useState } from "react"
 import { markdownRenderers as f0MarkdownRenderers } from "../markdownRenderers"
 
 export const AssistantMessage = ({
   message,
   isGenerating,
-  isCurrentMessage: _isCurrentMessage,
   isLoading,
   markdownTagRenderers,
   onCopy,
   onThumbsDown,
   onThumbsUp,
-  subComponent,
 }: AssistantMessageProps) => {
   const translations = useI18n()
   const containerRef = useRef<HTMLDivElement>(null)
   const [reactionValue, setReactionValue] = useState<"like" | "dislike" | null>(
     null
   )
-  useState(false)
   const [isHovered, setIsHovered] = useState(false)
   const timeoutRef = useRef<NodeJS.Timeout>()
-
   const handleMouseEnter = useCallback(() => {
     if (timeoutRef.current) {
       clearTimeout(timeoutRef.current)
     }
     setIsHovered(true)
   }, [])
-
   const handleMouseLeave = useCallback(() => {
     timeoutRef.current = setTimeout(() => {
       setIsHovered(false)
     }, 150)
   }, [])
+  const content = message?.content || ""
+  const subComponent = message?.generativeUI?.()
 
   return (
     <div
@@ -62,7 +58,7 @@ export const AssistantMessage = ({
         <>
           <div className="[&>div]:flex [&>div]:flex-col [&>div]:gap-1">
             <Markdown
-              content={message}
+              content={content}
               components={{ ...f0MarkdownRenderers, ...markdownTagRenderers }}
             />
           </div>
@@ -80,11 +76,11 @@ export const AssistantMessage = ({
               <div>
                 <CopyButton
                   variant="ghost"
-                  valueToCopy={message}
+                  valueToCopy={content}
                   disabled={isGenerating}
                   onCopy={(e) => {
                     e.currentTarget.blur()
-                    onCopy?.(message)
+                    onCopy?.(content)
                   }}
                 />
               </div>
@@ -98,9 +94,7 @@ export const AssistantMessage = ({
                   disabled={isGenerating}
                   onClick={(e) => {
                     setReactionValue((val) => (val === "like" ? null : "like"))
-                    // Blocked! new version of the sdk changes the type of the `message` variable so we can pass it directly
-                    // the current version does not provide enough information to create `TextMessage` correctly
-                    onThumbsUp?.({} as TextMessage)
+                    onThumbsUp?.(message)
                     e.currentTarget.blur()
                   }}
                 />
@@ -119,9 +113,7 @@ export const AssistantMessage = ({
                     setReactionValue((val) =>
                       val === "dislike" ? null : "dislike"
                     )
-                    // Blocked! new version of the sdk changes the type of the `message` variable so we can pass it directly
-                    // the current version does not provide enough information to create `TextMessage` correctly
-                    onThumbsDown?.({} as TextMessage)
+                    onThumbsDown?.(message)
                     e.currentTarget.blur()
                   }}
                 />