+++ packages/react/src/ui/scrollarea.tsx
@@ -95,8 +95,8 @@ const ScrollBar = forwardRef<
     className={cn(
       "group/scrollbar z-50 flex touch-none select-none p-[1px]",
       "transition-opacity data-[state=hidden]:pointer-events-none data-[state=visible]:pointer-events-auto data-[state=hidden]:opacity-0 data-[state=visible]:opacity-100",
-      orientation === "vertical" && "mr-[2px] h-full w-2",
-      orientation === "horizontal" && "mt-[2px] h-2 flex-col",
+      orientation === "vertical" && "h-full w-2",
+      orientation === "horizontal" && "h-2 flex-col",
       className
     )}
     {...props}